fix(ai): normalize batch result content
