+++ packages/ai/src/batch/batch-types.test-d.ts
@@ -28,6 +28,8 @@ import {
   type Experimental_TextBatchRequest as TextBatchRequest,
 } from '../index';
 import type { AsyncIterableStream } from '../util/async-iterable-stream';
+import type { ContentPart } from '../generate-text/content-part';
+import type { ToolSet } from '@ai-sdk/provider-utils';
 
 it('exposes typed Gateway async-job metadata', () => {
   expectTypeOf<
@@ -92,6 +94,12 @@ it('uses serializable response timestamps', () => {
   expectTypeOf<
     NonNullable<TextBatchGenerationResult['response']>['timestamp']
   >().toEqualTypeOf<string | undefined>();
+});
+
+it('exposes Core content in successful batch results', () => {
+  expectTypeOf<TextBatchGenerationResult['content']>().toEqualTypeOf<
+    Array<ContentPart<ToolSet>>
+  >();
 });
 
 it('flattens successful Core items while reusing provider status and errors', () => {
